Fixed some bugs
git-svn-id: https://roboticssrv.wtb.tue.nl/svn/ros/user/Janno/whole_body_planner@18098 3a3c2cfb-d3b9-409e-9e75-db8fcdbcaac6

+++ src/PlannerTopological.cpp
@@ -23,7 +23,7 @@ bool PlannerTopological::computeConstraints(const amigo_whole_body_controller::A
     /// Get a path from this goal
     constraints = connectivity_graph_.getPlan(current_state_,goal_constraint.goal_type);
 
-    ROS_INFO("Constraints: size = %i",constraints.size());
+    ROS_INFO("Constraints: size = %i",(int)constraints.size());
 
     /// Update the current state
     current_state_ = goal_constraint.goal_type;
